+++ src/main/java/team/creative/creativecore/common/util/math/vec/Vec1f.java
@@ -57,7 +57,7 @@ public class Vec1f extends VecNf<Vec1f> {
     
     @Override
     public void scale(double scale) {
-        this.x *= scale;
+        this.x *= (float) scale;
     }
     
     @Override
